+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 - Introduced the new `Welcome Screen` tab with the list of SAP Commerce projects [#1882](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1882)
 - Skip project import providers step and proceed to the SAP Project import wizard [#1890](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1890)
 - Display a SAP Commerce version [#1888](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1888)
-- Support deleting a recent project from the SAP Commerce Projects tab [1891](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1891)
+- Support deleting a recent project from the SAP Commerce Projects tab [#1891](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1891)
 
 ## [2026.0.8]